fix: fixing show hided comments finctions
