+++ js/bigpicture.js
@@ -1,60 +1,49 @@
 import {isEscapeKey} from './utils.js';
-
-const COMMENTS_PER_PORTION = 5;
-let commentsShown = 5;
+import { renderComments, showComments } from './comments.js';
 
 const bigPicture = document.querySelector('.big-picture'); // блок с полноразмерным изображением
 const bigPictureImgBlock = bigPicture.querySelector('.big-picture__img'); //изображение блока bigPicture
 const bigPictureImg = bigPictureImgBlock.querySelector('img'); //изображение блока bigPicture
 const bigPictureLikes = bigPicture.querySelector('.likes-count'); //Количество лайков likes
 const bigPictureCommentsCount = bigPicture.querySelector('.comments-count'); //Общее количество комментариев comments
-const bigPictureSocialCommentsCount = bigPicture.querySelector('.social__comment-count'); //Показанное количество комментариев comments
 const commentsLoader = document.querySelector('.comments-loader'); //кнопка загрузки комментариев
 const bigPictureCloseButton = bigPicture.querySelector('.big-picture__cancel'); //кнопка закрытия модалки
 
-const commentsList = document.querySelector('.social__comments');//Список комментариев <ul>
-const newCommentTemplate = document.querySelector('.social__comment');//шаблон комментария для наполнения
-const CommentsFragment = document.createDocumentFragment();//Фрагмент для наполнения и дальнейше вставки
+// const commentsList = document.querySelector('.social__comments');//Список комментариев <ul>
+// const newCommentTemplate = document.querySelector('.social__comment');//шаблон комментария для наполнения
+// const CommentsFragment = document.createDocumentFragment();//Фрагмент для наполнения и дальнейше вставки
 
 
 //функция отрисовки комментариев которая принимает в себя массив объектов
-const renderComments = (arrayOfComments) => {//arrayOfComments, каждый объект которого имеет ключи с именами avatar, message, name
-  commentsList.innerHTML = '';
+// const renderComments = (arrayOfComments) => {//arrayOfComments, каждый объект которого имеет ключи с именами avatar, message, name
+//   commentsList.innerHTML = '';
 
-  const slicedArray = arrayOfComments.slice(0, commentsShown);
-  if (arrayOfComments.length <= slicedArray.length) {
-    commentsLoader.classList.add('hidden');
-  }
+//   bigPictureCommentsCount.textContent = arrayOfComments.length;
 
-  bigPictureSocialCommentsCount.textContent = `${commentsShown} из`;
-  bigPictureCommentsCount.textContent = arrayOfComments.length;
+//   slicedArray.forEach(({avatar, message, name}) => {//при помощи деструктуризации передаем их в качетсве аргументов в forEach
+//     const comment = newCommentTemplate.cloneNode(true);//Копируем шаблон для наполнения на каждой итерации forEach
+//     const commentImg = comment.querySelector('img');//находим аватарку комментатора
 
-  slicedArray.forEach(({avatar, message, name}) => {//при помощи деструктуризации передаем их в качетсве аргументов в forEach
-    const comment = newCommentTemplate.cloneNode(true);//Копируем шаблон для наполнения на каждой итерации forEach
-    const commentImg = comment.querySelector('img');//находим аватарку комментатора
+//     commentImg.src = avatar;
+//     commentImg.alt = name;
+//     commentImg.width = '35';
+//     commentImg.height = '35';
 
-    commentImg.src = avatar;
-    commentImg.alt = name;
-    commentImg.width = '35';
-    commentImg.height = '35';
+//     comment.querySelector('.social__text').textContent = message;
 
-    comment.querySelector('.social__text').textContent = message;
-    CommentsFragment.append(comment);//запихиваем каждый шаблон в фрагмент на каждой итерации
-  });
-  commentsList.append(CommentsFragment);//запихиваем фрагмент с собранными шаблонами в список <ul> в верстке
-};
+//     CommentsFragment.append(comment);//запихиваем каждый шаблон в фрагмент на каждой итерации
+//   });
+//   commentsList.append(CommentsFragment);//запихиваем фрагмент с собранными шаблонами в список <ul> в верстке
+// };
 
 //отрисовка большого изображения, принимает в себя некий объект у которого есть ключи с именами url, likes, comments
 const renderBigPicture = ({url, likes, comments}) => {
   bigPictureImg.src = url;
   bigPictureLikes.textContent = likes;
   bigPictureCommentsCount.textContent = comments.length;
-
-  renderComments (comments);
-  commentsLoader.addEventListener('click', () => {
-    commentsShown += COMMENTS_PER_PORTION;
-
-    renderComments (comments);
+  showComments(renderComments(comments));
+  commentsLoader.addEventListener ('click', () => {
+    showComments(renderComments(comments));
   });
 };
 
@@ -70,7 +59,7 @@ const openBigPicture = (element) => {
     bigPicture.classList.add('hidden');//скрывает BigPicture добавив класс hidden
     document.documentElement.classList.remove('modal-open');//убирает класс modal-open на <body>
   });
-
+  //пусть при открытом состоянии при ниажатии на ESC окно закрывается
   document.addEventListener('keydown', (evt) => {
     if (isEscapeKey(evt)) {
       evt.preventDefault();
